pyproject.toml
[tool.poetry]
name = "securevision"
version = "0.1.0"
description = "SecureVision: Robust and Explainable Computer Vision Under Adversarial Attacks"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
pydantic = "^2.9.0"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.36"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
minio = "^7.2.9"
boto3 = "^1.35.18"
redis = "^5.0.8"
celery = "^5.4.0"
pendulum = "^3.0.0"
orjson = "^3.10.7"
pyyaml = "^6.0.2"
loguru = "^0.7.2"
requests = ">=2.25.1,<2.26.0"
# ML / CV
torch = "^2.3.1"
torchvision = "^0.18.1"
timm = "^1.0.9"
numpy = "^1.26.4"
opencv-python = "^4.10.0.84"
scikit-image = "^0.24.0"
scikit-learn = "^1.5.1"
pillow = "^10.4.0"
captum = "^0.7.0"
# Attacks
torchattacks = "^3.5.1"
foolbox = "^3.3.3"
robustbench = "^1.1"
# Reports
weasyprint = "^62.3"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
hypothesis = "^6.112.0"
httpx = "^0.27.2"
ruff = "^0.6.9"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
pre-commit = "^3.8.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
select = ["E","F","I","B","UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q --disable-warnings --maxfail=1 --cov=securevision --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
